[SOURCE-WEIGHT-002] Phase C dtype neutrality test fixes
Phase C complete - fixed dtype assertion issues in AT-SRC-001 tests.
Parser already implements dtype=None with torch.get_default_dtype() fallback.
Tests now use dtype from parser output instead of hardcoded float32.

Changes:
- tests/test_at_src_001.py: Use directions.dtype/weights.dtype in assertions
  instead of hardcoded float32 (3 locations)
- Preserves dtype neutrality when caller provides explicit dtype parameter

Test results: 10/10 passing (tests/test_at_src_001*.py)
- All AT-SRC-001 tests pass
- dtype regression tests pass (float32/float64/None)
- Expected warning for file λ vs CLI λ mismatch (per spec)

Refs: [TEST-SUITE-TRIAGE-001] Sprint 1.2, input.md Do Now
Next: update fix_plan with Attempt #17 artifacts

+++ tests/test_at_src_001.py
@@ -61,7 +61,8 @@ class TestAT_SRC_001_SourcefileAndWeighting:
 
             # Check directions are normalized (unit vectors)
             norms = torch.linalg.norm(directions, dim=1)
-            torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float32))
+            # Use dtype from parser output (defaults to torch.get_default_dtype())
+            torch.testing.assert_close(norms, torch.ones(2, dtype=directions.dtype))
 
             # Per spec-a-core.md:151-153, wavelength column is read but IGNORED.
             # CLI -lambda (default_wavelength_m) is the sole authoritative source.
@@ -146,7 +147,8 @@ class TestAT_SRC_001_SourcefileAndWeighting:
             )
 
             # Position [-15, 0, 0] normalized to unit vector: [-1, 0, 0]
-            expected_direction = torch.tensor([[-1.0, 0.0, 0.0]], dtype=torch.float32)
+            # Use dtype from parser output (respects beam_direction dtype, which is float64 here)
+            expected_direction = torch.tensor([[-1.0, 0.0, 0.0]], dtype=directions.dtype)
             torch.testing.assert_close(directions, expected_direction)
 
             # Per spec-a-core.md:151-153, wavelength uses CLI -lambda (default_wavelength_m)
@@ -179,7 +181,8 @@ class TestAT_SRC_001_SourcefileAndWeighting:
             torch.testing.assert_close(directions[0], -directions[1])
 
             # Both sources have weight 1.0 specified in the file
-            torch.testing.assert_close(weights, torch.ones(2, dtype=torch.float32))
+            # Use dtype from parser output (defaults to torch.get_default_dtype())
+            torch.testing.assert_close(weights, torch.ones(2, dtype=weights.dtype))
 
     def test_empty_sourcefile(self):
         """Test that empty sourcefile raises appropriate error."""
